[test] gathering apdu check tests
